% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-dim-models.R
\name{cmm2d_mde_ext}
\alias{cmm2d_mde_ext}
\title{Two-dimensional constant mixing model with the minimum-distance estimation routine: an extended variant}
\usage{
cmm2d_mde_ext(dat)
}
\arguments{
\item{dat}{An object returned by the function \code{refall2d}}
}
\value{
a list with four elements:
\itemize{
\item a tibble with the estimated coefficients and relevant statistical tests;
\item a matrix representing the model-predicted status allocation;
\item the index of dissimilarity between the observed and model-predicted status allocations;
\item the minimised distance (i.e., the Frobenius norm of the difference) between the observed and model-predicted status allocations
}

\code{cmm2d_mde_ext} implements an extension of \code{cmm2d_mde} in which the mixing coefficient on the secondary merit dimension is allowed to vary depending on whether status allocation with respect to the primary dimension is driven by lottery or meritocracy.
}
\description{
Two-dimensional constant mixing model with the minimum-distance estimation routine: an extended variant
}
\examples{
t1 <- refall2d(d = ks1985, dim1 = "degree", dim2 = "sex", status = "status", f = "freq")
cmm2d_mde_ext(t1)
}
