% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-dim-models.R
\name{cmm_mle}
\alias{cmm_mle}
\title{Constant mixing model with maximum-likelihood estimation routine}
\usage{
cmm_mle(dat)
}
\arguments{
\item{dat}{an object returned by \code{stat.alloc::mar()}, or a list with the following three elements: the observed status allocation, the meritocratic status allocation, and the lottery allocation.}
}
\value{
a list of three elements:
\itemize{
\item a data frame with the estimated mixing coefficient, its standard error and test statistics
\item a matrix representing the model-predicted status allocation table
\item the index of dissimilarity between observed and model-predicted status allocation tables. It is used as the model's goodness of fit measure.
}
}
\description{
The function estimates a mixing coefficient in the constant mixing model by means of maximum likelihood. \code{cmm} stands for "constant mixing model", while \code{mde} stands for "minimum distance estimation".
}
\examples{
ks1985 <- matrix(c(135, 52, 58, 12, 6, 27, 56, 126, 26, 7, 13, 51, 183, 98, 65, 3, 8, 30, 23, 2), ncol = 4, dimnames = list(rownames = paste0("E", 1:5), colnames = paste0("O", 1:4)))
ks1985_mar <- mar(x = ks1985)
cmm_mle(ks1985_mar)
}
