% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-dim-models.R
\name{dmm2d_mle}
\alias{dmm2d_mle}
\title{Two-dimensional differential mixing model with the maximum-likelihood estimation routine}
\usage{
dmm2d_mle(dat)
}
\arguments{
\item{dat}{An object returned by the function \code{refall2d}}
}
\value{
a list with five elements:
\itemize{
\item a tibble with the estimated coefficients and relevant statistical tests;
\item a named vector of adjustment proportions
\item the index of dissimilarity between the observed and model-predicted status allocations;
\item the minimised distance (i.e., the Frobenius norm of the difference) between the observed and model-predicted status allocations
\item a matrix representing the model-predicted status allocation;
}
}
\description{
Two-dimensional differential mixing model with the maximum-likelihood estimation routine
}
\examples{
t1 <- refall2d(xtabs(freq ~ origin + status, data = ks1985))
dmm2d_mde(t1)
}
