% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-dim-models.R
\name{dmm_mde}
\alias{dmm_mde}
\title{Differential mixing model with the minimum-distance estimation routine}
\usage{
dmm_mde(dat)
}
\arguments{
\item{dat}{an object returned by \code{stat.alloc::refall()}, or a list with the following three elements: the observed status allocation, the meritocratic status allocation, and the lottery allocation.}
}
\value{
a list of four elements:
\itemize{
\item a data frame with the estimated mixing coefficients, one coefficient per origin category, their standard error and associated test statistics
\item a named vector of \strong{adjustment proportions}, or conditional probabilities that an individual ends up in destination status \eqn{k} \strong{given that} the allocation is driven by lottery. Adjustment proportions are necessary to fit column marginals of the status allocation matrix
\item a matrix representing the model-predicted status allocation table
\item the index of dissimilarity between observed and model-predicted status allocation tables. It is used as the model's goodness of fit measure.
}
}
\description{
The function estimates mixing coefficients in the differential mixing model by minimising the distance between observed and model-predicted status allocations. \code{dmm} stands for "differential mixing model", while \code{mde} stands for "minimum distance estimation".
}
\examples{
data(ks1985)
t1 <- xtabs(freq ~ degree + status, data = ks1985)
t1_ref <- refall(x = t1)
dmm_mde(t1_ref)
}
