% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ks1985}
\alias{ks1985}
\title{1977 labour force participation by education, gender, and status}
\format{
\subsection{\code{ks1985}}{

A data frame with 40 rows and 4 columns:
\describe{
\item{edu}{Educational attainment classified into 5 categories:
\strong{E1} -- College, 4 years or more; \strong{E2} -- College, 1-3 years;
\strong{E3} -- High school, 4 years; \strong{E4} -- High school, 1-3 years;
\strong{E5} -- Elementary school, 8 years or less}
\item{sex}{Gender, coded \strong{M} for male and \strong{F} for female}
\item{status}{Occupational status classified into 4 categories:
\strong{S1} -- Professional, technical and kindred workers; Managers and administrators,
except farm; \strong{S2} -- Sales workers; Clerical and kindred workers; \strong{S3} -- Craft and
kindred workers; Operatives, except transport; Transport equipment operatives; Service
workers, except household; \strong{S4} -- laborers, except farm; Farm workers; Private household
workers}
\item{freq}{Frequencies (counts) of individuals in each joint category of \code{edu}, \code{sex}, and
\code{status}}
}
}
}
\source{
U.S. Department of Labor, Bureau of Labor Statistics. 1977. "Educational Attainment of
Workers", \href{https://www.google.pl/books/edition/Special_Labor_Force_Reports/t8VpM89u6XMC?hl=en&gbpv=1&pg=RA4-PP3&printsec=frontcover}{Special Labor Force Report 209, March 1977}
}
\usage{
ks1985
}
\description{
A subset of the 1977 report by the U.S. Department of Labour, Bureau of Labour Statistics, based on data in Table K in the Appendix of the original report (see the link below)
}
\keyword{datasets}
