% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar.R
\name{mar}
\alias{mar}
\title{Title}
\usage{
mar(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
A list of three matrices, each of the same order as \code{x}:
\itemize{
\item The first element of the list is the original status allocation matrix, \code{x}.
\item The second element of the list corresponds to a fully meritocratic status allocation, i.e., a status allocation which would be observed the allocation to destination statuses were driven solely by the principle of meritocracy, as defined by Krauze and Słomczyński (1985).
\item The third element of the list corresponds to a status allocation matrix which would be observed if the allocation to destination statuses were driven solely by lottery.
}
}
\description{
Title
}
\examples{
x <- matrix(c(10, 10, 0, 0, 0, 10, 20, 0, 0, 0, 20, 30), ncol = 4)
mar(x)
}
