% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refall.R
\name{refall2d}
\alias{refall2d}
\title{Two-dimensional model of status allocation}
\usage{
refall2d(d, dim1, dim2, status, f)
}
\arguments{
\item{d}{the data frame containing the data}

\item{dim1}{the name of the variable in \code{d} corresponding to the primary merit characteristic}

\item{dim2}{the name of the variable in \code{d} corresponding to the secondary merit characteristic}

\item{status}{the name of the variable in \code{d} corresponding to the destination status}

\item{f}{the name of the variable in \code{d} corresponding to the counts of unique combinations of the values of dim1, dim2, and status}
}
\value{
a list with five status allocation matrices:
\itemize{
\item actual (or observed);
\item meritocratic with respect to both dimensions;
\item meritocratic with respect to the primary and lottery with respect to the secondary dimension;
\item lottery with respect to the primary and meritocratic with respect to the secondary dimension; and
\item lottery with respect to both dimensions
}
}
\description{
Two-dimensional model of status allocation
}
\examples{
data(ks1985)
refall2d(d = ks1985, dim1 = "degree", dim2 = "sex", status = "status", f = "freq")
}
