% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-dim-models.R
\name{cmm_mde}
\alias{cmm_mde}
\title{Constant mixing model with the minimum-distance estimation routine}
\usage{
cmm_mde(dat)
}
\arguments{
\item{dat}{an object returned by \code{stat.alloc::refall()}, or a list with the following three elements: the observed status allocation, the meritocratic status allocation, and the lottery allocation.}
}
\value{
a list of three elements:
\itemize{
\item a data frame with the estimated mixing coefficient, its standard error and test statistics
\item a matrix representing the model-predicted status allocation table
\item the index of dissimilarity between observed and model-predicted status allocation tables. It is used as the model's goodness of fit measure.
}
}
\description{
The function estimates a mixing coefficient in the constant mixing model by minimising the distance between observed and model-predicted status allocations. \code{cmm} stands for "constant mixing model", while \code{mde} stands for "minimum distance estimation".
}
\examples{
data(ks1985)
t1 <- xtabs(freq ~ degree + status, data = ks1985)
t1_ref <- refall(x = t1)
cmm_mde(t1_ref)
}
